# api/internal/runtime/static/openapi.yaml
openapi: 3.0.3
info:
  title: NAAP Analytics API
  version: "1"
  description: |
    Read-only analytics API for the Livepeer AI network.
    Backed by ClickHouse materialized views fed from two Kafka topics:
    `network_events` (daydream org) and `streaming_events` (cloudspe org).

    **Rate limiting:** 100 req/s sustained, burst 200 per IP.
    Exceeded requests receive `429 Too Many Requests`.

    **Default time window:** last 24 hours when `start`/`end` are omitted.

    **WEI values** (payment amounts) are returned as decimal strings to avoid
    float64 precision loss (e.g. `"1234500000000000000"`).
servers:
  - url: https://naap-api.cloudspe.com
    description: Production
  - url: http://localhost:8000
    description: Local development
  - url: https://naap-api.livepeer.cloud
    description: Production
...
components:
  parameters:
    org:
      name: org
      in: query
      description: Filter to a single organisation. Omit for all orgs.
      schema:
        type: string
        enum: [daydream, cloudspe]
...
    stream_id:
      name: stream_id
      in: query
      schema:
        type: string
...
  schemas:
...
    NetworkSummary:
      type: object
      properties:
        Org:
          type: string
          example: ""
        SnapshotTime:
          type: string
          format: date-time
        TotalRegistered:
          type: integer
          format: int64
        TotalActive:
          type: integer
          format: int64
        ActiveThresholdMinutes:
          type: integer
...
  responses:
    429:
      description: Rate limit exceeded.
      headers:
        Retry-After:
          schema:
            type: integer
          description: Seconds to wait before retrying.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    500:
      description: Internal server error.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
...
paths:
...
  /v1/net/summary:
    get:
      tags: [Network]
      summary: Network summary
      description: Total and active orchestrator counts at the time of the last capabilities snapshot.
      operationId: getNetworkSummary
      parameters:
        - $ref: "#/components/parameters/org"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NetworkSummary"
        "429":
          $ref: "#/components/responses/429"
        "500":
          $ref: "#/components/responses/500"
  /v1/net/orchestrators:
    get:
      tags: [Network]
      summary: List orchestrators
      description: Paginated list of orchestrators ordered by last seen descending.
      operationId: listOrchestrators
      parameters:
        - $ref: "#/components/parameters/org"
        - $ref: "#/components/parameters/active_only"
        - $ref: "#/components/parameters/limit"
        - $ref: "#/components/parameters/offset"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Orchestrator"
        "429":
          $ref: "#/components/responses/429"
        "500":
          $ref: "#/components/responses/500"
  /v1/net/orchestrators/{address}:
    get:
      tags: [Network]
      summary: Orchestrator profile
      description: Full profile for a single orchestrator including 7-day reliability stats and raw capabilities JSON.
      operationId: getOrchProfile
      parameters:
        - $ref: "#/components/parameters/address_path"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrchProfile"
        "404":
          $ref: "#/components/responses/404"
        "429":
          $ref: "#/components/responses/429"
        "500":
          $ref: "#/components/responses/500"
  /v1/net/gpu:
    get:
      tags: [Network]
      summary: GPU summary
      description: Total GPU count and VRAM aggregated by hardware model.
      operationId: getGPUSummary
      parameters:
        - $ref: "#/components/parameters/org"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GPUSummary"
        "429":
          $ref: "#/components/responses/429"
        "500":
          $ref: "#/components/responses/500"
  /v1/net/models:
    get:
      tags: [Network]
      summary: Model availability
      description: Available AI models per pipeline, including warm orch count and price range.
      operationId: listModels
      parameters:
        - $ref: "#/components/parameters/org"
        - $ref: "#/components/parameters/pipeline"
        - $ref: "#/components/parameters/limit"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ModelAvailability"
        "429":
          $ref: "#/components/responses/429"
        "500":
          $ref: "#/components/responses/500"

  # ── Streams ────────────────────────────────────────────────────────────────
  /v1/streams/active:
    get:
      tags: [Streams]
      summary: Active streams
      description: Count of currently active streams broken down by org, pipeline, and state. A stream is considered active if it was seen within the last 60 seconds.
      operationId: getActiveStreams
      parameters:
        - $ref: "#/components/parameters/org"
        - $ref: "#/components/parameters/pipeline"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActiveStreamsSummary"
        "429":
          $ref: "#/components/responses/429"
        "500":
          $ref: "#/components/responses/500"
  /v1/streams/summary:
    get:
      tags: [Streams]
      summary: Stream activity summary
      description: Aggregate stream lifecycle metrics over the requested time window.
      operationId: getStreamSummary
      parameters:
        - $ref: "#/components/parameters/org"
        - $ref: "#/components/parameters/pipeline"
        - $ref: "#/components/parameters/start"
        - $ref: "#/components/parameters/end"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StreamSummary"
        "429":
          $ref: "#/components/responses/429"
        "500":
          $ref: "#/components/responses/500"
  /v1/streams/history:
    get:
      tags: [Streams]
      summary: Stream activity history
      description: Hourly time series of stream counts over the requested window.
      operationId: listStreamHistory
      parameters:
        - $ref: "#/components/parameters/org"
        - $ref: "#/components/parameters/pipeline"
        - $ref: "#/components/parameters/start"
        - $ref: "#/components/parameters/end"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StreamBucket"
        "429":
          $ref: "#/components/responses/429"
        "500":
          $ref: "#/components/responses/500"

  # ── Performance ────────────────────────────────────────────────────────────
  /v1/perf/fps:
    get:
      tags: [Performance]
      summary: FPS summary
      description: |
        Inference and input FPS percentiles (p5/p50/p99) broken down by pipeline
        and by orchestrator. Samples with fps == 0 are excluded.
        Computed from raw events (not aggregates) for percentile accuracy.
      operationId: getFPSSummary
      parameters:
        - $ref: "#/components/parameters/org"
        - $ref: "#/components/parameters/pipeline"
        - $ref: "#/components/parameters/orch_address"
        - $ref: "#/components/parameters/start"
        - $ref: "#/components/parameters/end"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FPSSummary"
        "429":
          $ref: "#/components/responses/429"
        "500":
          $ref: "#/components/responses/500"
  /v1/perf/fps/history:
    get:
      tags: [Performance]
      summary: FPS history
      description: Hourly average inference and input FPS for charting. Buckets flagged as `Degraded` when avg FPS drops ≥20% from the prior bucket.
      operationId: listFPSHistory
      parameters:
        - $ref: "#/components/parameters/org"
        - $ref: "#/components/parameters/pipeline"
        - $ref: "#/components/parameters/orch_address"
        - $ref: "#/components/parameters/start"
        - $ref: "#/components/parameters/end"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FPSBucket"
        "429":
          $ref: "#/components/responses/429"
        "500":
          $ref: "#/components/responses/500"
  /v1/perf/latency:
    get:
      tags: [Performance]
      summary: Orchestrator discovery latency
      description: Avg, p50, p95, p99 discovery latency per orchestrator, computed from raw `discovery_results` events.
      operationId: getLatencySummary
      parameters:
        - $ref: "#/components/parameters/org"
        - $ref: "#/components/parameters/start"
        - $ref: "#/components/parameters/end"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LatencySummary"
        "429":
          $ref: "#/components/responses/429"
        "500":
          $ref: "#/components/responses/500"
  /v1/perf/webrtc:
    get:
      tags: [Performance]
      summary: WebRTC quality
      description: Aggregated WebRTC video/audio jitter, packet loss, and connection quality distribution.
      operationId: getWebRTCQuality
      parameters:
        - $ref: "#/components/parameters/org"
        - $ref: "#/components/parameters/stream_id"
        - $ref: "#/components/parameters/start"
        - $ref: "#/components/parameters/end"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WebRTCQuality"
        "429":
          $ref: "#/components/responses/429"
        "500":
          $ref: "#/components/responses/500"

  # ── Payments ───────────────────────────────────────────────────────────────
  /v1/payments/summary:
    get:
      tags: [Payments]
      summary: Payment summary
      description: Total payments in WEI, event count, and unique orchestrators over the window. Broken down by org.
      operationId: getPaymentSummary
      parameters:
        - $ref: "#/components/parameters/org"
        - $ref: "#/components/parameters/start"
        - $ref: "#/components/parameters/end"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentSummary"
        "429":
          $ref: "#/components/responses/429"
        "500":
          $ref: "#/components/responses/500"
  /v1/payments/history:
    get:
      tags: [Payments]
      summary: Payment history
      description: Hourly payment totals for charting.
      operationId: listPaymentHistory
      parameters:
        - $ref: "#/components/parameters/org"
        - $ref: "#/components/parameters/start"
        - $ref: "#/components/parameters/end"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PaymentBucket"
        "429":
          $ref: "#/components/responses/429"
        "500":
          $ref: "#/components/responses/500"
  /v1/payments/by-pipeline:
    get:
      tags: [Payments]
      summary: Payments by pipeline
      description: Total payments and average price per pixel for each pipeline.
      operationId: listPaymentsByPipeline
      parameters:
        - $ref: "#/components/parameters/org"
        - $ref: "#/components/parameters/pipeline"
        - $ref: "#/components/parameters/start"
        - $ref: "#/components/parameters/end"
        - $ref: "#/components/parameters/limit"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PipelinePayment"
        "429":
          $ref: "#/components/responses/429"
        "500":
          $ref: "#/components/responses/500"
  /v1/payments/by-orch:
    get:
      tags: [Payments]
      summary: Payments by orchestrator
      description: Total payments received per orchestrator, ordered by total descending.
      operationId: listPaymentsByOrch
      parameters:
        - $ref: "#/components/parameters/org"
        - $ref: "#/components/parameters/orch_address"
        - $ref: "#/components/parameters/start"
        - $ref: "#/components/parameters/end"
        - $ref: "#/components/parameters/limit"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OrchPayment"
        "429":
          $ref: "#/components/responses/429"
        "500":
          $ref: "#/components/responses/500"

  # ── Reliability ────────────────────────────────────────────────────────────
  /v1/reliability/summary:
    get:
      tags: [Reliability]
      summary: Reliability summary
      description: Network-wide stream success rate, failure rates, and failure breakdown over the window.
      operationId: getReliabilitySummary
      parameters:
        - $ref: "#/components/parameters/org"
        - $ref: "#/components/parameters/start"
        - $ref: "#/components/parameters/end"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReliabilitySummary"
        "429":
          $ref: "#/components/responses/429"
        "500":
          $ref: "#/components/responses/500"
  /v1/reliability/history:
    get:
      tags: [Reliability]
      summary: Reliability history
      description: Hourly success and failure rates for charting. Rates are null when the sample count is below the minimum threshold (5 streams).
      operationId: listReliabilityHistory
      parameters:
        - $ref: "#/components/parameters/org"
        - $ref: "#/components/parameters/start"
        - $ref: "#/components/parameters/end"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ReliabilityBucket"
        "429":
          $ref: "#/components/responses/429"
        "500":
          $ref: "#/components/responses/500"
  /v1/reliability/orchs:
    get:
      tags: [Reliability]
      summary: Orchestrator reliability
      description: Per-orchestrator degraded, restart, and error rates over the window.
      operationId: listOrchReliability
      parameters:
        - $ref: "#/components/parameters/org"
        - $ref: "#/components/parameters/orch_address"
        - $ref: "#/components/parameters/start"
        - $ref: "#/components/parameters/end"
        - $ref: "#/components/parameters/limit"
        - $ref: "#/components/parameters/offset"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OrchReliability"
        "429":
          $ref: "#/components/responses/429"
        "500":
          $ref: "#/components/responses/500"
  /v1/failures:
    get:
      tags: [Reliability]
      summary: Failure events
      description: Raw failure event log, newest first. Filterable by type.
      operationId: listFailures
      parameters:
        - $ref: "#/components/parameters/org"
        - $ref: "#/components/parameters/failure_type"
        - $ref: "#/components/parameters/start"
        - $ref: "#/components/parameters/end"
        - $ref: "#/components/parameters/limit"
        - $ref: "#/components/parameters/offset"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FailureEvent"
        "429":
          $ref: "#/components/responses/429"
        "500":
          $ref: "#/components/responses/500"

  # ── Leaderboard ────────────────────────────────────────────────────────────
  /v1/leaderboard:
    get:
      tags: [Leaderboard]
      summary: Orchestrator leaderboard
      description: |
        Orchestrators ranked by composite score.
        Scoring weights: FPS 30%, reliability (1−degradedRate) 30%,
        volume 20%, latency (inverted) 20%.
        All four dimensions are min-max normalised across the result set,
        then scaled to 0–100. Only orchestrators with ≥10 streams handled
        in the window are included.
      operationId: getLeaderboard
      parameters:
        - $ref: "#/components/parameters/org"
        - $ref: "#/components/parameters/start"
        - $ref: "#/components/parameters/end"
        - $ref: "#/components/parameters/limit"
        - $ref: "#/components/parameters/offset"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LeaderboardEntry"
        "429":
          $ref: "#/components/responses/429"
        "500":
          $ref: "#/components/responses/500"
  /v1/leaderboard/{address}:
    get:
      tags: [Leaderboard]
      summary: Leaderboard orchestrator profile
      description: Full profile for a single orchestrator including 7-day reliability stats and raw capabilities JSON.
      operationId: getLeaderboardProfile
      parameters:
        - $ref: "#/components/parameters/address_path"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrchProfile"
        "404":
          $ref: "#/components/responses/404"
        "429":
          $ref: "#/components/responses/429"
        "500":
          $ref: "#/components/responses/500"

  # ── Gateways (R7) ────────────────────────────────────────────────────────
  /v1/net/gateways:
    get:
      summary: List gateways
      description: Returns all known gateways with active stream counts.
      operationId: listGateways
      tags: [Gateways]
      parameters:
        - $ref: "#/components/parameters/org"
        - $ref: "#/components/parameters/limit"
        - $ref: "#/components/parameters/offset"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Gateway" }
        "429": { $ref: "#/components/responses/TooManyRequests" }
  /v1/net/gateways/{address}:
    get:
      summary: Gateway profile
      description: Returns stats and metadata for one gateway.
      operationId: getGatewayProfile
      tags: [Gateways]
      parameters:
        - $ref: "#/components/parameters/address_path"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/GatewayProfile" }
        "404": { $ref: "#/components/responses/NotFound" }
        "429": { $ref: "#/components/responses/TooManyRequests" }
  /v1/net/gateways/{address}/orchs:
    get:
      summary: Gateway orchestrators
      description: Returns orchestrators that handled traffic for a given gateway.
      operationId: listGatewayOrchestrators
      tags: [Gateways]
      parameters:
        - $ref: "#/components/parameters/address_path"
        - $ref: "#/components/parameters/limit"
        - $ref: "#/components/parameters/offset"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/GatewayOrch" }
        "429": { $ref: "#/components/responses/TooManyRequests" }

  # ── Stream Extensions (R8) ───────────────────────────────────────────────
  /v1/streams/samples:
    get:
      summary: Stream status samples
      description: Returns recent per-stream telemetry samples from the last 30 days.
      operationId: listStreamSamples
      tags: [Streams]
      parameters:
        - $ref: "#/components/parameters/org"
        - $ref: "#/components/parameters/start"
        - $ref: "#/components/parameters/end"
        - $ref: "#/components/parameters/pipeline"
        - $ref: "#/components/parameters/orch_address"
        - $ref: "#/components/parameters/stream_id"
        - $ref: "#/components/parameters/limit"
        - $ref: "#/components/parameters/offset"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/StreamStatusSample" }
        "429": { $ref: "#/components/responses/TooManyRequests" }
  /v1/streams/attribution:
    get:
      summary: Stream attribution summary
      description: Returns the fraction of stream samples that were successfully attributed to an orchestrator.
      operationId: getAttributionSummary
      tags: [Streams]
      parameters:
        - $ref: "#/components/parameters/org"
        - $ref: "#/components/parameters/start"
        - $ref: "#/components/parameters/end"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AttributionSummary" }
        "429": { $ref: "#/components/responses/TooManyRequests" }
  /v1/streams/{stream_id}:
    get:
      summary: Stream detail
      description: Returns the full state and sample history for one stream.
      operationId: getStreamDetail
      tags: [Streams]
      parameters:
        - name: stream_id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/StreamDetail" }
        "404": { $ref: "#/components/responses/NotFound" }
        "429": { $ref: "#/components/responses/TooManyRequests" }

  # ── E2E Latency (R9) ────────────────────────────────────────────────────
  /v1/perf/e2e-latency:
    get:
      summary: E2E latency summary
      description: Returns end-to-end latency percentiles from stream status samples.
      operationId: getE2ELatencySummary
      tags: [Performance]
      parameters:
        - $ref: "#/components/parameters/org"
        - $ref: "#/components/parameters/start"
        - $ref: "#/components/parameters/end"
        - $ref: "#/components/parameters/pipeline"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/E2ELatencySummary" }
        "429": { $ref: "#/components/responses/TooManyRequests" }
  /v1/perf/e2e-latency/history:
    get:
      summary: E2E latency history
      description: Returns hourly E2E latency buckets (p50, p95, avg).
      operationId: listE2ELatencyHistory
      tags: [Performance]
      parameters:
        - $ref: "#/components/parameters/org"
        - $ref: "#/components/parameters/start"
        - $ref: "#/components/parameters/end"
        - $ref: "#/components/parameters/pipeline"
        - $ref: "#/components/parameters/limit"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/E2ELatencyBucket" }
        "429": { $ref: "#/components/responses/TooManyRequests" }

  # ── Pipelines (R10) ─────────────────────────────────────────────────────
  /v1/pipelines:
    get:
      summary: List pipelines
      description: Returns a cross-cutting summary per pipeline combining streams, FPS, and payment data.
      operationId: listPipelines
      tags: [Pipelines]
      parameters:
        - $ref: "#/components/parameters/org"
        - $ref: "#/components/parameters/start"
        - $ref: "#/components/parameters/end"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/PipelineSummary" }
        "429": { $ref: "#/components/responses/TooManyRequests" }
  /v1/pipelines/{pipeline}:
    get:
      summary: Pipeline detail
      description: Returns detailed stats for one pipeline including model breakdown.
      operationId: getPipelineDetail
      tags: [Pipelines]
      parameters:
        - name: pipeline
          in: path
          required: true
          description: Pipeline name (e.g. `text-to-image`).
          schema: { type: string }
        - $ref: "#/components/parameters/org"
        - $ref: "#/components/parameters/start"
        - $ref: "#/components/parameters/end"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PipelineDetail" }
        "404": { $ref: "#/components/responses/NotFound" }
        "429": { $ref: "#/components/responses/TooManyRequests" }

  # ── Pricing (R11) ───────────────────────────────────────────────────────
  /v1/net/pricing:
    get:
      summary: List orchestrator pricing
      description: Returns flat pricing rows for all active orchestrators.
      operationId: listPricing
      tags: [Pricing]
      parameters:
        - $ref: "#/components/parameters/org"
        - $ref: "#/components/parameters/pipeline"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/OrchPricingEntry" }
        "429": { $ref: "#/components/responses/TooManyRequests" }
  /v1/net/pricing/{address}:
    get:
      summary: Orchestrator pricing profile
      description: Returns structured pipeline/model pricing for one orchestrator.
      operationId: getOrchPricingProfile
      tags: [Pricing]
      parameters:
        - $ref: "#/components/parameters/address_path"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/OrchPricingProfile" }
        "404": { $ref: "#/components/responses/NotFound" }
        "429": { $ref: "#/components/responses/TooManyRequests" }

  # ── Model Performance (R12) ──────────────────────────────────────────────
  /v1/perf/by-model:
    get:
      summary: Performance by AI model
      description: Returns FPS performance broken down by AI model, joined from GPU inventory.
      operationId: listModelPerformance
      tags: [Performance]
      parameters:
        - $ref: "#/components/parameters/org"
        - $ref: "#/components/parameters/start"
        - $ref: "#/components/parameters/end"
        - $ref: "#/components/parameters/pipeline"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/ModelPerformance" }
        "429": { $ref: "#/components/responses/TooManyRequests" }
...